refactor template request
